requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-http
pandas
requests_ntlm
SQLAlchemy
beautifulsoup4
workalendar
gspread
google-auth-oauthlib